Add elapsed-time to Application and set up /app/stop

+++ local-kube/src/main/java/fr/umlv/localkube/services/ApplicationService.java
@@ -7,6 +7,6 @@ import java.util.Optional;
 
 public interface ApplicationService {
     List<Application> getAll();
-    Optional<Application> findById(long id);
+    Optional<Application> findById(int id);
     void save(Application app);
 }
